daybook-be/pyproject.toml
[project]
name = "daybook-be"
version = "0.1"
requires-python = ">=3.10"

dependencies = [
    "fastapi ~= 0.105.0",
    "uvicorn",
    "httpx ~= 0.25.2",
    "pydantic-settings ~= 2.1.0",
    "lagom ~= 2.6.0",
    "python-dotenv ~= 1.0.0",
    "logtail-python ~= 0.2.9",
    "PyJWT",
    "PyDrive2",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "google-api-python-client",
    "pygsheets"
]
[project.optional-dependencies]
dev = [
    "ruff ~= 0.1.8",
    "black ~= 23.12.0",
    "mypy ~= 1.7.1",
    "mutmut ~= 2.4.4",
    "coverage ~= 7.3.3",
]

[tool.ruff]
ignore = [
    "E501", # line too long
    "B008", # do not perform function call ... in argument defaults - needed for Depends/Header of fastapi
]

[tool.ruff.per-file-ignores]
"tests/**" = ["E402"]  # ignore imports that are not at the top of the file, which we allow in tests

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
# these packages don't have type hints, stop mypy complaining about them
module = ["logtail"]
ignore_missing_imports = true
